tmpfs: add user_xattr mount option
We want user xattrs on tmpfs to support flatpak running as an
unprivileged user on overlayfs whose upper fs is tmpfs, a situation
which only arises in our pristine image boots. So, for our purposes, a
mount flag is okay: we can set it at the same time as setting up the
overlays.

For the reasons discussed at
<http://www.spinics.net/lists/linux-mm/msg109982.html>, this simplistic
implementation of user xattrs can't simply be made available to
everyone.

https://phabricator.endlessm.com/T13817

Signed-off-by: Will Thompson <wjt@endlessm.com>
[jprvita: Option parsing re-written for Linux 5.4]
Signed-off-by: João Paulo Rechi Vita <jprvita@endlessm.com>